Fix out of order PUBACK and PUBLISH handling
The current implementation uses task notification to wait for
acknowledgements and incoming publishes and therefore, cannot
differentiate between acknowledgements and incoming publishes. This
causes a crash when an incoming PUBLISH is received before PUBACK for
the previous outgoing publish.

This commit uses an event group to wait for acknowledgements and
incoming publishes. This allows us to differentiate between
acknowledgements and incoming publishes by waiting for different bits.

This was reported here - https://github.com/FreeRTOS/iot-reference-esp32c3/issues/63.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>